--- core/home/templates/home/base.html
@@ -1,9 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-	<title></title>
-</head>
-<body>
-	{% block content %}{% endblock content %}
-</body>
-</html>
