+++ core/home/urls.py
@@ -1,6 +1,9 @@
 from django.urls import path
-from . import views
+from .views import home_view, MaterialDetailView
+from django.conf.urls import url
 
 urlpatterns = [
-    path('', views.home, name='home'),
+    url(r'^$', home_view, name='home'),
+    url(r'^material/(?P<pk>\d+)$', MaterialDetailView.as_view(), name='material-detail'),
+
 ]
