+++ core/registration/views.py
@@ -14,23 +14,6 @@ from search.models import Material
 from operator import attrgetter
 
 
-def home_view(request):
-    context = {}
-
-    query = ""
-    if request.GET:
-        query = request.GET['q']
-        context['query'] = str(query)
-
-    if query == "":
-        print("Empty request")
-        material_list = sorted(Material.objects.all(), key=attrgetter('date_publication'), reverse=True)
-    else:
-        material_list = sorted(get_material_queryset(query), key=attrgetter('date_publication'), reverse=True)
-    context['material_list'] = material_list
-    return render(request, 'material_list.html', context)
-
-
 def signup_view(request):
     form = SignUpForm(request.POST)
     if form.is_valid():
