+++ core/search/templates/search/material_list.html
@@ -1,6 +1,20 @@
+
+
+
 {% extends "base_generic.html" %}
 
 {% block content %}
+<div class="search-bar mt-sm-2 mr-2">
+    <form method="get">
+        <input type="text" class='form-control' name="q" id='id_q' placeholder="Search...">
+
+    </form>
+
+</div>
+<script type="text/javascript">
+    document.getElementById("id_q").value = "{{query}}"
+</script>
+
     <h1>material List</h1>
 
     {% if material_list %}
