+++ core/search/views.py
@@ -1,8 +1,9 @@
 from django.shortcuts import render
 from .models import Material
 from django.db.models import Q
-# Create your views here.
 
+
+# Create your views here.
 
 
 # Word by word finding matches
@@ -14,9 +15,9 @@ def get_material_queryset(query=None):
     for q in queries:
         materials = Material.objects.filter(
             (Q(title__icontains=q) |
-            Q(author__icontains=q) |
-            Q(file_name__icontains=q) |
-            Q(author__icontains=q)) &
+             Q(author__icontains=q) |
+             Q(file_name__icontains=q) |
+             Q(author__icontains=q)) &
             Q(visibility__icontains='1')
         ).distinct()
 
@@ -24,4 +25,3 @@ def get_material_queryset(query=None):
             queryset.append(material)
 
     return list(set(queryset))
-
