+++ core/templates/base_generic.html
@@ -24,12 +24,12 @@
             <ul class="sidebar-nav">
                 <li><a href="{% url 'home' %}">Home</a></li>
                 {% if user.is_authenticated %}
-                    <li>User: {{ user.get_username }}</li>
-                    <li><a href="{% url 'logout'%}?next={{request.path}}">Logout</a></li>
-                    <li><a href="{% url 'password_change'%}?next={{request.path}}">Change password</a></li>
+                <li>User: {{ user.get_username }}</li>
+                <li><a href="{% url 'logout'%}?next={{request.path}}">Logout</a></li>
+                <li><a href="{% url 'password_change'%}?next={{request.path}}">Change password</a></li>
                 {% else %}
-                    <li><a href="{% url 'login'%}?next={{request.path}}">Login</a></li>
-                    <li><a href="{% url 'password_reset'%}?next={{request.path}}">Reset password</a></li>
+                <li><a href="{% url 'login'%}?next={{request.path}}">Login</a></li>
+                <li><a href="{% url 'password_reset'%}?next={{request.path}}">Reset password</a></li>
                 {% endif %}
                 <li><a href="{% url 'signup' %}?next={{request.path}}">Registrate</a></li>
                 <li><a href="/admin/">Admin panel</a></li>
@@ -37,15 +37,6 @@
             {% endblock %}
         </div>
         <div class="col-sm-10 ">
-
-            <div class="search-bar mt-sm-2 mr-2">
-                <form method="get">
-                    <input type="text" class='form-control' name="q" id = 'id_q' placeholder="Search...">
-
-                </form>
-
-            </div>
-
 
 
             {% block content %}{% endblock %}
@@ -69,9 +60,6 @@
         </div>
     </div>
 
-    <script type="text/javascript">
-        document.getElementById("id_q").value = "{{query}}"
-    </script>
 
 </div>
 </body>
