--- core/templates/index.html
@@ -1,107 +1,0 @@
-{% extends "base_generic.html" %}
-{% load static %}
-{% block content %}
-<h1>Electronic Library Home Page</h1>
-
-<p>Welcome to <em>Electronic Library</em>, a very basic web-app, created with love by Daniil, Arina, Maxim, Igor and
-    Maxim</p>
-
-<section class="most-search spad">
-        <div class="container">
-            <div class="row">
-                <div class="col-lg-12">
-                    <div class="section-title">
-                        <h2>Library Content</h2>
-                    </div>
-                </div>
-            </div>
-            <div class="row">
-                <div class="col-lg-12">
-                    <div class="most__search__tab">
-                        <ul class="nav nav-tabs" role="tablist">
-                            <li class="nav-item">
-                                <a class="nav-link active" data-toggle="tab" href="#tabs-1" role="tab">
-                                    Lectures
-                                </a>
-                            </li>
-                            <li class="nav-item">
-                                <a class="nav-link" data-toggle="tab" href="#tabs-2" role="tab">
-                                    Assignments
-                                </a>
-                            </li>
-                            <li class="nav-item">
-                                <a class="nav-link" data-toggle="tab" href="#tabs-3" role="tab">
-                                    Books
-                                </a>
-                            </li>
-                            <li class="nav-item">
-                                <a class="nav-link" data-toggle="tab" href="#tabs-4" role="tab">
-                                    Tutorials
-                                </a>
-                            </li>
-                        </ul>
-                    </div>
-                    <div class="tab-content">
-                        <div class="tab-pane active" id="tabs-1" role="tabpanel">
-                            <div class="row">
-                                {% for material in Lecture %}
-                                    <div class="col-lg-4 col-md-6">
-                                        <div class="listing__item">
-                                            <h6>Title: {{ material.title }}</h6>
-                                            <h6>Author: {{ material.author }}</h6>
-                                            <h6>Type: {{ material.material_type }}</h6>
-                                            <h6>Status: {{ material.status }}</h6>
-                                        </div>
-                                    </div>
-                                {% endfor %}
-                            </div>
-                        </div>
-                        <div class="tab-pane" id="tabs-2" role="tabpanel">
-                            <div class="row">
-                                {% for material in Assignment %}
-                                    <div class="col-lg-4 col-md-6">
-                                        <div class="listing__item">
-                                            <h6>Title: {{ material.title }}</h6>
-                                            <h6>Author: {{ material.author }}</h6>
-                                            <h6>Type: {{ material.material_type }}</h6>
-                                            <h6>Status: {{ material.status }}</h6>
-                                        </div>
-                                    </div>
-                                {% endfor %}
-                            </div>
-                        </div>
-                        <div class="tab-pane" id="tabs-3" role="tabpanel">
-                            <div class="row">
-                                {% for material in Book %}
-                                    <div class="col-lg-4 col-md-6">
-                                        <div class="listing__item">
-                                            <h6>Title: {{ material.title }}</h6>
-                                            <h6>Author: {{ material.author }}</h6>
-                                            <h6>Type: {{ material.material_type }}</h6>
-                                            <h6>Status: {{ material.status }}</h6>
-                                        </div>
-                                    </div>
-                                {% endfor %}
-                            </div>
-                        </div>
-                        <div class="tab-pane" id="tabs-4" role="tabpanel">
-                            <div class="row">
-                                {% for material in Lecture %}
-                                    <div class="col-lg-4 col-md-6">
-                                        <div class="listing__item">
-                                            <h6>Title: {{ material.title }}</h6>
-                                            <h6>Author: {{ material.author }}</h6>
-                                            <h6>Type: {{ material.material_type }}</h6>
-                                            <h6>Status: {{ material.status }}</h6>
-                                        </div>
-                                    </div>
-                                {% endfor %}
-                            </div>
-                        </div>
-                    </div>
-                </div>
-            </div>
-        </div>
-    </section>
-
-{% endblock %}
